move settings icon onto observationList screen
